libroach: fix the windows build
The second declaration of `rocksDBLog` was not marked `weak` which broke
the build with the Windows linker. Move the weak go symbols into
`godefs.{cc,h}`.

Release note (bug fix): Fix Windows release builds.
